Move import & clone logic to RepackImporter
No actual logic code changes, only required
changes in order to compile (forward to specific
interfaces for required functionality)

Also, had to make public some classes in order
to use interfaces (sad that C# doesn't even look
at interface's access modifier. If it's internal, internal
properties should be enough, but it wants public anyway)
