+++ ILRepack/ILRepack.csproj
@@ -86,6 +86,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ReferenceFixator.cs" />
     <Compile Include="ReflectionHelper.cs" />
+    <Compile Include="RepackImporter.cs" />
     <Compile Include="RepackLogger.cs" />
     <Compile Include="RepackOptions.cs" />
     <Compile Include="RepackAssemblyResolver.cs" />
